# apps/miniapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Bot Manager API",
    "description": "REST API behind the Telegram Mini App admin panel.\n\nAuthenticate by posting the WebApp `initData` string to `/api/auth/telegram`,\nthen send the returned token as `Authorization: Bearer <token>` on every other\nrequest. Tokens are short-lived; renew them with `/api/auth/refresh` rather than\nre-posting `initData`, which Telegram only lets you use once.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/platform/stats": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Platform-wide totals for the operator dashboard",
        "operationId": "getPlatformStats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStats"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramSession": []
          }
        ]
      }
    },
    "/api/platform/dashboard": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Platform dashboard with a selectable reporting window",
        "operationId": "getPlatformDashboard",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformDashboard"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/users": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Search and paginate all known Telegram users",
        "operationId": "listPlatformUsers",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "banned_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Banned Only"
            }
          },
          {
            "name": "admins_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Admins Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUserPage"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/users/{tg_user_id}": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Get a user and the chats they administer",
        "operationId": "getPlatformUser",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "tg_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUserDetail"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/users/{tg_user_id}/subscription": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Grant a subscription to one of a user's chats",
        "operationId": "grantPlatformSubscription",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "tg_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformSubscriptionGrant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/payments": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List platform payments for support and reconciliation",
        "operationId": "listPlatformPayments",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentProvider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "tg_user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tg User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformPaymentPage"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/settings": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Read deployment capabilities and operator settings",
        "operationId": "getPlatformSettings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformSettings"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OperationResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "type": "object",
        "title": "OperationResult"
      },
...
      "PlatformStats": {
        "properties": {
          "total_chats": {
            "type": "integer",
            "title": "Total Chats"
          },
          "active_chats": {
            "type": "integer",
            "title": "Active Chats"
          },
          "chats_by_plan": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Chats By Plan"
          },
          "revenue_stars": {
            "type": "integer",
            "title": "Revenue Stars"
          },
          "revenue_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Revenue Usd"
          },
          "global_bans": {
            "type": "integer",
            "title": "Global Bans"
          }
        },
        "type": "object",
        "required": [
          "total_chats",
          "active_chats",
          "chats_by_plan",
          "revenue_stars",
          "revenue_usd",
          "global_bans"
        ],
        "title": "PlatformStats"
      },
...
      "PlatformUserPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlatformUser"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "PlatformUserPage"
      },
...
    },
    "securitySchemes": {
      "TelegramSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}